+++ RELEASES.md
@@ -4,6 +4,7 @@
 - 'Default' option for _data source_ removal: now all _data sources_ are set at 'Auto' by default. We reccomend to leave them at 'Auto' unless you have a specific need: the best source will automatically be chosen based on the availability of dependencies, performance and reliability
 - New source selection for Memory Usage: now you can choose between GTop and /proc (GTop is the default if available)
 - New source selection for Storage Usage: now you can choose between GTop and /proc (GTop is the default if available)
+- New source selection for Network IO: now you can choose between GTop and /proc (GTop is the default if available)
 
 ### Bug fixes
 - Light Theme readability improvements [[#53](https://github.com/AstraExt/astra-monitor/issues/53)]